requirements.txt
pandas==1.5.3
streamlit==1.31.1
fastapi==0.109.2
pymongo ==4.6.1
weaviate-client==3.26.2
langchain-community==0.0.20
requests==2.31.0
uvicorn==0.27.1
langchain_openai==0.0.6
python-dotenv==1.0.1
langchain==0.1.7
pyodbc==5.1.0
openpyxl==3.1.2
pandasai== 2.0.6
pymysql==1.1.0